taipei-day-trip/requirements.txt
Flask~=2.2.2
mysql-connector-python
flask-restful
flask-bcrypt
flask-marshmallow
python-dotenv
pyjwt
requests
